Delete + Update Tavolo completati

+++ src/main/java/it/pokeronline/service/tavolo/TavoloServiceImpl.java
@@ -1,9 +1,12 @@
 package it.pokeronline.service.tavolo;
 
 import java.util.List;
+import java.util.Set;
 
 import javax.persistence.EntityManager;
+import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
+import javax.persistence.TypedQuery;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +14,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import it.pokeronline.model.tavolo.Tavolo;
+import it.pokeronline.model.user.User;
 import it.pokeronline.repository.tavolo.TavoloRepository;
 
 @Component
@@ -35,9 +39,17 @@ public class TavoloServiceImpl implements TavoloService {
 		return tavoloRepository.findById(id).orElse(null);
 	}
 
-	@Override
-	public void aggiorna(Tavolo tavoloInstance) {
+	@Transactional
+	public boolean aggiorna(Tavolo tavoloInstance) throws Exception {
+		boolean result = false;
+		Tavolo tavoloDaDB = this.caricaSingoloTavolo(tavoloInstance.getId());
+		if(tavoloDaDB.getUsers().size() == 0) {
 		tavoloRepository.save(tavoloInstance);
+			result = true;
+		} else {
+			result = false;
+		}
+		return result;
 	}
 
 	@Override
@@ -45,9 +57,17 @@ public class TavoloServiceImpl implements TavoloService {
 		tavoloRepository.save(tavoloInstance);
 	}
 
-	@Override
-	public void rimuovi(Tavolo tavoloInstance) {
+	@Transactional
+	public boolean rimuovi(Tavolo tavoloInstance) throws Exception {
+		boolean result = false;
+		Tavolo tavoloDaDB = this.caricaSingoloTavolo(tavoloInstance.getId());
+		if(tavoloDaDB.getUsers().size() == 0) {
 		tavoloRepository.delete(tavoloInstance);
+			result = true;
+		} else {
+			result = false;
+		}
+		return result;
 	}
 
 	@Override
